create new components for easy development
